+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Roact Changelog
 
 ## Unreleased Changes
+
+## [1.3.0](https://github.com/Roblox/roact/releases/tag/v1.3.0) (May 5th, 2020)
 * Added Contexts, which enables easy handling of items that are provided and consumed throughout the tree.
 
 ## [1.2.0](https://github.com/Roblox/roact/releases/tag/v1.2.0) (September 6th, 2019)